# SmartPowerElectricAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SmartPowerElectricAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Trabajador/create": {
      "post": {
        "tags": [
          "Trabajador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TrabajadorDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Trabajador/{id}": {
      "delete": {
        "tags": [
          "Trabajador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Trabajador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TrabajadorDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "TrabajadorDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "especialidad": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "seguridadSocial": {
          "type": "string"
        },
        "fechaInicioContrato": {
          "type": "string"
        },
        "fechaFinContrato": {
          "type": "string"
        },
        "cobroxHora": {
          "format": "double",
          "type": "number"
        },
        "numeroCuenta": {
          "type": "string"
        },
        "enrutamiento": {
          "type": "string"
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
  }
}